updated state.gd with init() and run() methods
states no longer run off of _process(), they now use update() which does general state tree stuff, and run() which the individual state implements.

There is also an init() function which enters all the initial states. Passes an empty array to _args which may not be the cleanest thing, but I can't think of a solution right now